Fix signout action
